sandbox/requirements.old.txt
apipkg==1.5
appdirs==1.4.3
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==2.1.1
bootstrap-admin==0.4.3
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.5.4
decorator==4.4.2
defusedxml==0.5.0
distlib==0.3.0
Django==1.11.29
django-allauth==0.32.0
django-appconf==1.0.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.8
django-extra-views==0.11.0
django-haystack==2.7.0
django-hosts==4.0
django-imagekit==4.0.1
django-js-asset==1.2.2
django-phonenumber-field==2.0.1
django-redis==4.9.0
django-tables2==1.21.2
django-treebeard==4.3.1
django-webtest==1.9.4
django-widget-tweaks==1.4.5
djangorestframework==3.8.2
easy-thumbnails==2.5
emoji==0.4.5
entrypoints==0.3
execnet==1.7.1
factory-boy==2.12.0
Faker==4.0.1
filelock==3.0.12
flake8==3.7.9
flake8-blind-except==0.1.1
flake8-debugger==3.2.1
freezegun==0.3.15
futures==3.1.1
goslate==1.5.1
gunicorn==19.7.1
html5lib==1.0b10
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.0.2
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jieba==0.39
Markdown==2.6.8
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==2.0.3
olefile==0.44
packaging==20.1
parso==0.6.2
pexpect==4.8.0
phonenumbers==8.11.4
pickleshare==0.7.5
pilkit==2.0
Pillow==4.2.1
pluggy==0.13.1
polib==1.1.0
prompt-toolkit==3.0.3
psycopg2==2.7.7
psycopg2-binary==2.8.4
ptyprocess==0.6.0
purl==1.5
py==1.8.1
pycodestyle==2.5.0
pycountry==19.8.18
pyflakes==2.1.1
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.4.6
pyprof2calltree==1.4.4
pysolr==3.8.1
pytest==4.4.2
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-forked==1.1.3
pytest-xdist==1.27.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sorl-thumbnail==12.4.1
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
tox==3.8.6
traitlets==4.3.3
urllib3==1.22
user-agent==0.1.9
uWSGI==2.0.18
virtualenv==20.0.7
waitress==1.4.3
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.6
WebTest==2.0.34
Werkzeug==1.0.0
whitenoise==5.0.1
Whoosh==2.7.4
zipp==3.0.0
